Move lab node class to common
